mysite/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.10
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
